converters/requirements.txt
PyYAML
redis
paho-mqtt
scipy
